dap: Fully extract template parameter prompts
